pyproject.toml
[project]
name = "phytonet"
version = "0.1.0"
description = "Phytoplankton classification using deep learning"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
keywords = [
  "phytoplankton",
  "classification",
  "deep-learning",
  "computer-vision",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "timm>=0.9.0",
  "pillow>=9.0.0",
  "tqdm>=4.64.0",
  "pathlib>=1.0.0",
  "pandas>=2.0.0",
  "pyarrow>=10.0.0",
]

[project.scripts]
phytonet-train = "phytonet.cli:train_cli"
phytonet-predict = "phytonet.cli:predict_cli"

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "black>=22.0.0",
  "isort>=5.10.0",
  "flake8>=4.0.0",
  "mypy>=1.0.0",
  "pre-commit>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/phytonet"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
